refactoring icrc wallet
