+++ src/frontend/src/wallet.tsx
@@ -16,7 +16,7 @@ import {
 } from "./common";
 import { Principal } from "@dfinity/principal";
 import { CANISTER_ID } from "./env";
-import { Icrc1TokensWallet } from "./icrc1-tokens-wallet";
+import { Icrc1TokensWallet } from "./tokens_wallet";
 
 const coldWalletFunctionalityAvailable = window.ic && window.ic.plug;
